Update linter configs
