--- gui/.eslintignore
@@ -1,4 +1,0 @@
-.eslintrc.js
-gulpfile.js
-init.js
-tasks/*.js
