--- gui/.prettierrc.yml
@@ -1,11 +1,0 @@
----
-# .prettierrc.yml
-# see: https://prettier.io/docs/en/options.html
-printWidth: 100
-semi: true
-singleQuote: true
-trailingComma: all
-bracketSpacing: true
-jsxBracketSameLine: true
-arrowParens: always
-proseWrap: always
