+++ gui/prettier.config.mjs
@@ -1,0 +1,7 @@
+// see: https://prettier.io/docs/en/options.html
+export default {
+  printWidth: 100,
+  singleQuote: true,
+  bracketSameLine: true,
+  proseWrap: 'always',
+};
